requirements.txt
litellm
fastapi
uvicorn
pyyaml
requests
